Completing the Book details endpoint
The endpoint should return the cover image, title, author name, average review rating, the list of reviews for the book, and (if the user is logged in) the checkout-out status, and (again if the user is logged in) any of the user's reviews should be indicated.  To support this functionality, anonymous users are enabled, and the response varies based on whether the user is logged in or anonymous.

Resolves:  GL-2